+++ src/main/java/kcs/funding/fundingboost/domain/service/utils/FundingConst.java
@@ -1,4 +1,4 @@
-package kcs.funding.fundingboost.domain.service;
+package kcs.funding.fundingboost.domain.service.utils;
 
 public interface FundingConst {
     int EXTEND_DEADLINE = 14;
